# entityBackend/api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Issuer & Service Provider backend
host: localhost:10010
basePath: /v1/entity
...
schemes:
  - http
  - https
consumes:
  - application/octet-stream
  - application/json
  - text/plain
  - text/html
produces:
  - application/octet-stream
  - application/json
  - text/plain
  - text/html
paths:
...
  /credential/db:
    x-swagger-router-controller: user.controller
    get:
      description: Get available credentials catalog from an issuer
      operationId: getCredentialFromDB
      parameters:
        - name: authToken
          in: query
          type: string
          description: Auth Token to validate in tha call to backend
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/createCredentialResponse"
        "400":
          description: Bad Request
          schema:
            $ref: "#/definitions/ErrorResponse"
        "404":
          description: Address not found
          schema:
            $ref: "#/definitions/ErrorResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
  /user:
    x-swagger-router-controller: user.controller
    post:
      description: Create a new User in the database
      # used as the method name of the controller
      operationId: addUser
      parameters:
        - name: body
          in: body
          description: Necessary data to create a new user in database
          required: false
          schema: 
            $ref: "#/definitions/userData"
      responses:
        "200":
          description: User creation success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/userCreatedSuccess"
        # responses may fall through to errors
        "400":
          description: Bad Request
          schema:
            $ref: "#/definitions/ErrorResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    put:
      description: Update use data in the database
      # used as the method name of the controller
      operationId: updateUser
      parameters:
        - name: id
          in: query
          description: Id of the user to update
          required: true
          type: string
        - name: body
          in: body
          description: Necessary data to update an user in database
          required: false
          schema: 
            $ref: "#/definitions/userData"
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/userData"
        # responses may fall through to errors
        "400":
          description: Bad Request
          schema:
            $ref: "#/definitions/ErrorResponse"
        "404":
          description: User not found
          schema:
            $ref: "#/definitions/ErrorResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    get:
      description: Get data saved in the database
      operationId: getUser
      parameters:
        - name: userID
          in: query
          type: string
          description: Id of the user
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/userInfo"
        "400":
          description: Bad Request
          schema:
            $ref: "#/definitions/ErrorResponse"
        "404":
          description: Address not found
          schema:
            $ref: "#/definitions/ErrorResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /user/checkAuth:
    x-swagger-router-controller: user.controller
    get:
      description: check if user is authenticated
      operationId: checkUserAuth
      parameters:
        - name: authToken
          in: query
          type: string
          description: tokoen to check
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/userAuthenticated"
        "400":
          description: Bad Request
          schema:
            $ref: "#/definitions/ErrorResponse"
        "404":
          description: Address not found
          schema:
            $ref: "#/definitions/ErrorResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /user/credential/{identityDID}:
    x-swagger-router-controller: user.controller
    get:
      description: Get available credentials catalog from an issuer
      operationId: getCredentialIdentityCatalog
      parameters:
        - name: identityDID
          in: path
          type: string
          description: Identity to find the credential catalog
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/credentialCatalogList"
        "400":
          description: Bad Request
          schema:
            $ref: "#/definitions/ErrorResponse"
        "404":
          description: Address not found
          schema:
            $ref: "#/definitions/ErrorResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:
...
  userAuthenticated:
    properties:
      status:
        type: boolean
...
  userData:
    properties:
      username:
        type: string
      name:
        type: string
      surname:
        type: string
      email:
        type: string
      address:
        type: string
      password:
        type: string
      vinculated:
        type: boolean
      did:
        type: string
  userCreatedSuccess:
    required:
      - message
    properties:
      message:
        type: string
  userInfo:
    properties:
      data:
        type: object
# End definitions for user
...
  credentialCatalogList:
    properties: 
      list:
        $ref: "#/definitions/credentialCatalog"
  credentialCatalog:
    properties:
      credentialName:
        type: string
      credentialType:
        type: string
      credentialValue:
        type: string
      credentialFormat:
        type: object
...
  createCredentialResponse:
    properties:
      signedCredentials:
        type: array
        items:
          type: string
...
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string